+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add commands to query the word with modification to command palette.
 - Add menus to the editor context menu.
 - Refactoring: Add CscopeConfig and CscopeLog class.
-- Refactoring: Split CscopePosition class.
+- Refactoring: Split CscopePosition class and move some functionality.
 
 ## [0.0.4]
 ### Added